Refactor download dialog
