+++ src/lib/store.ts
@@ -17,6 +17,7 @@ export const theme = storable('theme', 'purple')
 export const showHand = storable('showHand', true)
 export const bomMultiplier = storable('bomMultiplier', '2')
 export const stiltsMsg = storable('stiltsMsg', true)
+export const modelName = storable('modelName', 'cosmotyl')
 export const developer = storable('developer', false)
 export const showTiming = andcondition(developer, storable('developer.timing', false))
 export const noWall = andcondition(developer, storable('developer.hideWall', false))
